Create Sum of even numbers
kuty and pappu plays a coding game. kuty give a set of numbers to pappu to find the even numbers in the set and sum that. Help pappu to solve the game and to win the diary milk.

Input Format:

Input consists of n+1 integers. The first integer corresponds to ‘n’ , the size of the array. The next ‘n’ integers correspond to the elements in the array. Assume that the maximum value of n is 15.

Output Format:

Refer sample output for details.

Sample Input :

5

2

3

6

8

-1

Sample Output 1:

The sum of the even numbers in the array is 16